Move analytics to platform.py and fix review findings
Migrate the anonymous analytics callbacks out of callbacks/hub.py into
callbacks/platform.py so they survive removal of the hub directory. They sit in
platform.py's dict outside the _api_key gate: gating them on an API key would
have silently reduced analytics to Platform users only. Sentry needs no
migration, it lives in utils/__init__.py and never referenced hub.

Events no longer depend on SETTINGS["hub"], which disappears with hub. The
documented sync setting in Events.enabled is now the only gate, so users who set
hub=False but left sync=True will send events again.

Pass the predictor rather than pre-read attributes, so every attribute access
happens inside the guard. An earlier revision read validator.model in the
callback and crashed real training runs with AttributeError, because
BaseValidator never assigns self.model. Val is reverted to on_val_start;
benchmark runs are already captured, since benchmark() calls predict() per
format at benchmarks.py:212 and each of those now reports format, arch, imgsz,
quantize, hardware and speeds.

Review fixes:
- accept the unscaled or the literal config stem, so per-scale families such as
  yolov10n resolve instead of collapsing to "custom"
- compare against a set of shipped stems instead of passing a user-controlled
  stem to rglob, where "[y]olo11n.yaml" matched yolo11.yaml and would have
  transmitted a private filename
- read quantize from export metadata when cfg.quantize is None, so int8 exports
  are not grouped with fp32
- report the resolved predictor.imgsz and the real dataset.bs instead of the
  requested values
- reset speed per run, so an embed run cannot report the previous run's latency
- revert the persistent GPU cache: the write is disk I/O that can log an error
  inside predict, and an index-keyed entry goes stale under CUDA_VISIBLE_DEVICES

Validated lineage on real fine-tunes: gen1, gen2 and gen3 descendants of
yolo26n.pt all report yolo26n, as does gen2 exported to ONNX, and a yolo11n-seg
fine-tune keeps its task suffix.
