+++ ultralytics/utils/callbacks/hub.py
@@ -5,7 +5,6 @@ from time import time
 
 from ultralytics.hub import HUB_WEB_ROOT, PREFIX, HUBTrainingSession
 from ultralytics.utils import LOGGER, RANK, SETTINGS
-from ultralytics.utils.events import events
 
 
 def on_pretrain_routine_start(trainer):
@@ -72,27 +71,6 @@ def on_train_end(trainer):
         LOGGER.info(f"{PREFIX}Done ✅\n{PREFIX}View model at {session.model_url} 🚀")
 
 
-def on_train_start(trainer):
-    """Run events on train start."""
-    events(trainer.args, trainer.device)
-
-
-def on_val_start(validator):
-    """Run events on validation start."""
-    if not validator.training:
-        events(validator.args, validator.device)
-
-
-def on_predict_end(predictor):
-    """Run events on predict end, once per-image speeds are known."""
-    events(predictor.args, predictor.device, model=predictor.model, speed=predictor.speed, n=predictor.seen)
-
-
-def on_export_start(exporter):
-    """Run events on export start."""
-    events(exporter.args, exporter.device)
-
-
 callbacks = (
     {
         "on_pretrain_routine_start": on_pretrain_routine_start,
@@ -100,10 +78,6 @@ callbacks = (
         "on_fit_epoch_end": on_fit_epoch_end,
         "on_model_save": on_model_save,
         "on_train_end": on_train_end,
-        "on_train_start": on_train_start,
-        "on_val_start": on_val_start,
-        "on_predict_end": on_predict_end,
-        "on_export_start": on_export_start,
     }
     if SETTINGS["hub"] is True
     else {}
