+++ ultralytics/utils/callbacks/platform.py
@@ -22,6 +22,7 @@ from ultralytics.utils import (
     Retry,
     colorstr,
 )
+from ultralytics.utils.events import events
 
 PREFIX = colorstr("Platform: ")
 PLATFORM_API_URL = os.getenv("PLATFORM_API_URL", f"{PLATFORM_URL}/api/webhooks")
@@ -598,14 +599,42 @@ def on_train_end(trainer):
     LOGGER.info(f"{PREFIX}View results at {url}")
 
 
-callbacks = (
-    {
-        "on_pretrain_routine_start": on_pretrain_routine_start,
-        "on_pretrain_routine_end": on_pretrain_routine_end,
-        "on_fit_epoch_end": on_fit_epoch_end,
-        "on_model_save": on_model_save,
-        "on_train_end": on_train_end,
-    }
-    if _api_key
-    else {}
-)
+def on_train_start(trainer):
+    """Run events on train start."""
+    events(trainer.args, trainer.device)
+
+
+def on_val_start(validator):
+    """Run events on validation start."""
+    if not validator.training:
+        events(validator.args, validator.device)
+
+
+def on_predict_end(predictor):
+    """Run events on predict end, once per-image speeds are known."""
+    events(predictor.args, predictor.device, predictor)
+
+
+def on_export_start(exporter):
+    """Run events on export start."""
+    events(exporter.args, exporter.device)
+
+
+callbacks = {
+    # Anonymous analytics, gated only by the documented sync setting inside Events
+    "on_train_start": on_train_start,
+    "on_val_start": on_val_start,
+    "on_predict_end": on_predict_end,
+    "on_export_start": on_export_start,
+    **(
+        {
+            "on_pretrain_routine_start": on_pretrain_routine_start,
+            "on_pretrain_routine_end": on_pretrain_routine_end,
+            "on_fit_epoch_end": on_fit_epoch_end,
+            "on_model_save": on_model_save,
+            "on_train_end": on_train_end,
+        }
+        if _api_key
+        else {}
+    ),
+}
